requirements.txt
# ─── Core ──────────────────────────────────────────────────
langchain>=0.2.0
langchain-huggingface>=0.0.3
langchain-community>=0.2.0

# ─── LLM ───────────────────────────────────────────────────
transformers>=4.40.0
torch>=2.2.0
accelerate>=0.28.0
bitsandbytes>=0.43.0          # 4-bit quantisation

# ─── Embeddings ────────────────────────────────────────────
sentence-transformers>=2.7.0

# ─── Vector Store ──────────────────────────────────────────
chromadb>=0.5.0

# ─── Sparse Retrieval ──────────────────────────────────────
rank_bm25>=0.2.2

# ─── Evaluation ────────────────────────────────────────────
ragas>=0.1.7
datasets>=2.19.0
pandas>=2.0.0

# ─── Scientific Embeddings (optional) ──────────────────────
# Uncomment to use SciBERT for technical queries:
# scibert @ git+https://github.com/allenai/scibert.git

# ─── PDF Support (optional) ────────────────────────────────
pypdf>=4.0.0

# ─── BEIR benchmark loader (for SciDocs-RAG) ───────────────
beir>=2.0.0

# ─── Config & Utils ────────────────────────────────────────
pyyaml>=6.0
tqdm>=4.66.0

# ─── Testing ───────────────────────────────────────────────
pytest>=8.0.0
pytest-cov>=5.0.0
